Fix issue with LazyImage component getting stuck in English fallback state
When the LazyImage component fails to load a localized card image it sets the fallback state to true and attempts the load the English version of the card. This state is not being reset when the card property changes which causes all subsequent image loads to be in English as well.